+++ robe-common/src/main/java/io/robe/common/service/search/Relation.java
@@ -1,0 +1,12 @@
+package io.robe.common.service.search;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+/**
+ * Created by kamilbukum on 20/01/2017.
+ */
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Relation {
+    String name();
+}