+++ robe-hibernate/src/main/java/io/robe/hibernate/query/api/criteria/cache/EntityMeta.java
@@ -8,10 +8,14 @@ import java.util.Map;
 public class EntityMeta {
     private final String identityName;
     private final Map<String, FieldMeta> fieldMap;
+    private final Map<String, String> fieldRelationMap;
+    private final Map<String, FieldMeta> relationMap;
 
-    public EntityMeta(String identityName, Map<String, FieldMeta> fieldMap) {
+    public EntityMeta(String identityName, Map<String, FieldMeta> fieldMap, Map<String, FieldMeta> relationMap, Map<String, String> fieldRelationMap) {
         this.identityName = identityName;
         this.fieldMap = fieldMap;
+        this.relationMap = relationMap;
+        this.fieldRelationMap = fieldRelationMap;
     }
 
     public String getIdentityName() {
@@ -21,4 +25,13 @@ public class EntityMeta {
     public Map<String, FieldMeta> getFieldMap() {
         return fieldMap;
     }
+
+    public Map<String, String> getFieldRelationMap() {
+        return fieldRelationMap;
+    }
+
+    public Map<String, FieldMeta> getRelationMap(){
+        return this.relationMap;
+    }
+
 }