+++ robe-hibernate/src/main/java/io/robe/hibernate/query/api/criteria/cache/FieldMeta.java
@@ -1,32 +1,51 @@
 package io.robe.hibernate.query.api.criteria.cache;
 
+import java.lang.reflect.Field;
+
 /**
  * Created by kamilbukum on 12/01/2017.
  */
 public class FieldMeta {
     private final boolean searchIgnore;
-    private final Class<?> type;
+    private final boolean hasRelation;
+    private final boolean isTransient;
+    private final Field field;
     private final FieldReference reference;
 
-    public FieldMeta(Class<?> type, boolean searchIgnore){
-        this(type, null, searchIgnore);
+    public FieldMeta(Field field, boolean isTransient, boolean searchIgnore, boolean hasRelation){
+        this(field, null, isTransient, searchIgnore, hasRelation);
     }
 
-    public FieldMeta(Class<?> type, FieldReference reference, boolean searchIgnore){
-        this.searchIgnore = searchIgnore;
-        this.type = type;
+    public FieldMeta(Field field, FieldReference reference, boolean isTransient, boolean searchIgnore, boolean hasRelation){
+        this.field = field;
         this.reference = reference;
-    }
+        this.isTransient = isTransient;
+        this.searchIgnore = searchIgnore;
+        this.hasRelation = hasRelation;
 
-    public boolean isSearchIgnore() {
-        return searchIgnore;
     }
 
     public FieldReference getReference() {
         return reference;
     }
 
+    public Field getField() {
+        return field;
+    }
+
     public Class<?> getType() {
-        return type;
+        return field.getType();
+    }
+
+    public boolean isTransient() {
+        return isTransient;
+    }
+
+    public boolean isSearchIgnore() {
+        return searchIgnore;
+    }
+
+    public boolean hasRelation(){
+        return this.hasRelation;
     }
 }